+++ common-core/change-log.md
@@ -10,6 +10,7 @@
 7. CollectionUtils加入map(K,V)方法
 8. Lang加入map(K,V)，get(Optional|Supplier)，is[Not]Null(Optional|Supplier)方法
 9. ChainMap加入set(K,V), set(Map)方法（功能和putChain,putAllChain方法一样）
+9. ClassUtils加入getPrimitiveType方法，forName方法支持基本类型(primitive type)
 
 # 1.14.0 2025-02-13